+++ src/main/java/com/zeylin/eventsourcingwithcqrs/exceptions/PayloadSerializationException.java
@@ -1,0 +1,9 @@
+package com.zeylin.eventsourcingwithcqrs.exceptions;
+
+public class PayloadSerializationException extends Exception {
+
+    public PayloadSerializationException(String s) {
+        super(s);
+    }
+
+}