+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 # Changelog
 All notable changes to this project will be documented in this file. The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [1.0.0-SNAPSHOT] - 2020-03-23
+### Added
+- Aggregate for showtimes
+- Base event and base command
+- Events and commands to create and update showtimes
+- Command and query controllers
+- Command and event handlers
+- Command dispatcher, aggregate utility
+
 ## [0.1.0-SNAPSHOT] - 2020-03-10
 ### Added
 - Initial project setup